fix: Attribute not found error
