+++ src/models/__init__.py
@@ -1,6 +1,7 @@
 from src.models.user import User
 from src.models.user_usage import UserUsage
 from src.models.user_setting import UserSetting
-from src.models.search_result import SearchResult
-from src.models.username import Username
-from src.models.search_usernames import search_usernames
+from src.models.search_result import SearchResult, Username, search_usernames
+
+# from src.models.username import Username
+# from src.models.search_usernames import search_usernames
