+++ src/models/username.py
@@ -10,22 +10,22 @@ if TYPE_CHECKING:
     from src.models.search_result import SearchResult
 
 
-class Username(db.Model):
-    id = mapped_column(
-        String(length=50), primary_key=True, default=lambda: uuid.uuid4().hex
-    )
-    username: Mapped[str] = mapped_column(
-        String(length=100), nullable=False, unique=True
-    )
-    type: Mapped[str] = mapped_column(String(length=10), nullable=False)
+# class Username(db.Model):
+#     id = mapped_column(
+#         String(length=50), primary_key=True, default=lambda: uuid.uuid4().hex
+#     )
+#     username: Mapped[str] = mapped_column(
+#         String(length=100), nullable=False, unique=True
+#     )
+#     type: Mapped[str] = mapped_column(String(length=10), nullable=False)
 
-    search_result: Mapped[List["SearchResult"]] = relationship(
-        secondary=search_usernames, back_populates="username"
-    )
+#     search_result: Mapped[List["SearchResult"]] = relationship(
+#         secondary=search_usernames, back_populates="username"
+#     )
 
-    def serialize(self):
-        return {
-            "id": self.id,
-            "username": self.username,
-            "type": self.type,
-        }
+#     def serialize(self):
+#         return {
+#             "id": self.id,
+#             "username": self.username,
+#             "type": self.type,
+#         }
